use previous seconds' worth of data to identify characters at race start
This renders the previous "test suite" useless, since it only tested identifying
characters from a single frame.  In development I've replaced this with a suite
that looks at several frames.  These frames are too large to check in, but it's
easy to generate them from the original video files using extract-frames.sh.