internal refactor to make server look like client
Previously, there was a bunch of complex logic in the client for translating the
output of "kartvid" into useful race records.  This has been moved to the
server for both simplicity and efficiency.  (Previously, a lot of code dealt
with both the janky internal format and the nice external one.  Now, everything
on the client and much of the server only needs to deal with the nice format.)
A subsequent change could make kartvid's format require a little less massaging
(e.g., change "position" to "rank", as was done for the "nice" format).

This change also replaces the embedded video for each race video with a per-race
video instead.

This change also adds status and progress for uploading videos.