Fix "keithings"
After the big interface change, parseKartvid was still populating segment.end
instead of segment.vend, which confused the higher-level stuff.  I've also
changed the kartlytics server to reprocess the events file whenever it comes up
so that we don't have to rerun everything just to change the way that gets
processed.